requirement.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
albumentations==1.0.3
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
blis==0.7.11
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.0.4
cloudpathlib==0.18.1
colorama==0.4.6
confection==0.1.5
contourpy==1.2.1
cycler==0.12.1
cymem==2.0.8
Cython==3.0.10
dataclasses==0.6
decorator==5.1.1
docker-pycreds==0.4.0
docstring_parser==0.16
effdet==0.2.4
efficientnet_pytorch==0.7.1
exceptiongroup==1.2.1
executing==2.0.1
fastai==2.5.6
fastcore==1.3.29
fastdownload==0.0.7
fastprogress==1.0.3
filelock==3.15.4
fire==0.6.0
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
ftfy==6.2.0
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.31.0
google-auth-oauthlib==0.4.6
grpcio==1.64.1
huggingface-hub==0.23.4
icedata==0.5.1
icevision==0.12.0
idna==3.7
imageio==2.34.2
importlib_metadata==8.0.0
importlib_resources==6.4.0
iopath==0.1.10
ipykernel==5.5.6
ipython==8.18.1
ipython-genutils==0.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonargparse==4.31.0
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
lightning-bolts==0.7.0
lightning-flash==0.8.2
lightning-utilities==0.11.3.post0
loguru==0.7.2
marisa-trie==1.2.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl-fft @ file:///C:/b/abs_19i1y8ykas/croot/mkl_fft_1695058226480/work
mkl-random @ file:///C:/b/abs_edwkj1_o69/croot/mkl_random_1695059866750/work
mkl-service==2.4.0
multidict==6.0.5
munch==4.0.0
murmurhash==1.0.10
networkx==3.2.1
nose==1.3.7
numpy @ file:///C:/b/abs_c1ywpu18ar/croot/numpy_and_numpy_base_1708638681471/work/dist/numpy-1.26.4-cp39-cp39-win_amd64.whl#sha256=f284633067fde4bb078a09085a1c9e99a3218c129f5312b7b3d76c9340442a25
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.2
parso==0.8.4
Pillow==8.4.0
platformdirs==4.2.2
portalocker==2.10.0
preshed==3.0.9
pretrainedmodels==0.7.4
prompt_toolkit==3.0.47
protobuf==3.19.6
psutil==6.0.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybboxes==0.1.5
pycocotools==2.0.8
pydantic==2.8.2
pydantic_core==2.20.1
pyDeprecate==0.3.2
Pygments==2.18.0
pyparsing==3.1.2
pystiche==1.0.1
python-dateutil==2.9.0.post0
pytorch-lightning==1.9.5
pytz==2024.1
PyWavelets==1.6.0
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.3
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
resnest==0.0.6b20230402
rich==13.7.1
rsa==4.9
safetensors==0.4.3
sahi==0.10.8
scikit-image==0.19.3
scikit-learn==1.5.1
scipy==1.13.1
seaborn==0.13.2
segmentation-models-pytorch==0.3.3
sentry-sdk==2.7.1
setproctitle==1.3.3
shapely==2.0.4
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
tabulate==0.9.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.4.0
terminaltables==3.1.10
thinc==8.2.5
threadpoolctl==3.5.0
tifffile==2024.7.2
timm==0.9.2
torch==1.10.1
torchaudio==0.10.1
torchmetrics==0.10.3
torchvision==0.11.2
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typer==0.12.3
typeshed_client==2.5.1
typing_extensions @ file:///C:/b/abs_0as9mdbkfl/croot/typing_extensions_1715268906610/work
tzdata==2024.1
urllib3==2.2.2
wandb==0.17.4
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
Werkzeug==3.0.3
win32-setctime==1.1.0
wrapt==1.16.0
yacs==0.1.8
yarl==1.9.4
yolov5-icevision==6.0.0
zipp==3.19.2
